submit

+++ to do list/index.php
@@ -1,10 +1,19 @@
 <?php
 require "model/DBConnection.php";
-require "model/CustomerDB.php";
-require "model/Customer.php";
-require "controller/CustomerController.php";
+require "model/TaskDB.php";
+require "model/Task.php";
+require "controller/TaskController.php";
+require "model/UserDB.php";
+require "model/User.php";
+require "controller/UserController.php";
 
-use \Controller\CustomerController;
+use \Controller\UserController;
+
+use \Controller\TaskController;
+
+// session_start();
+// require 'vendor/autoload.php';
+
 ?>
 <!DOCTYPE html>
 <html lang="en">
@@ -13,38 +22,40 @@ use \Controller\CustomerController;
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Thêm mới khách hàng</title>
+    <title>Thêm mới Task</title>
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css"
         integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO"
         crossorigin="anonymous">
 </head>
 
 <body>
-    <div class="container">
-        <div class="navbar navbar-default">
-            <a class="navbar-brand" href="index.php">Quản lý khách hàng</a>
-        </div>
+    <?php
 
-        <?php
-        $controller = new CustomerController();
-        $page = isset($_REQUEST['page']) ? $_REQUEST['page'] : null;
+    $page = $_GET['page'] ?? 'login';
 
-        switch ($page) {
-            case 'add':
-                $controller->add();
-                break;
-            case 'delete':
-                $controller->delete();
-                break;
-            case 'edit':
-                $controller->edit();
-                break;
-            default:
-                $controller->index();
-                break;
-        }
-        ?>
-    </div>
+    if ($page === 'login') {
+        $controller = new UserController();
+        $controller->login();
+    } elseif ($page === 'register') {
+        $controller = new UserController();
+        $controller->register();
+    } elseif ($page === 'todos') {
+        $controller = new TaskController();
+        $controller->index();
+    } elseif ($page === 'add') {
+        $controller = new TaskController();
+        $controller->add();
+    } elseif ($page === 'delete') {
+        $controller = new TaskController();
+        $controller->delete();
+    } elseif ($page === 'edit') {
+        $controller = new TaskController();
+        $controller->edit();
+    } else {
+        echo "Page not found!";
+    } ?>
+
+
 </body>
 <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
     integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
